refactor(profile): route fullscreen profile to the new page, drop old view
Rewire _showFullscreenProfile to push FullscreenProfilePage (Task 10)
instead of the old private _FullscreenProfilePage/_FullscreenProfilePageState,
which are deleted along with their private helpers. Extract
calculateSacNormalizationFactor into a shared utils file to avoid a
mutual import between dive_detail_page.dart and fullscreen_profile_page.dart.
Remove the four orphaned diveLog_detail_fullscreen_* l10n keys.
