+++ lib/l10n/arb/app_ar.arb
@@ -1116,10 +1116,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{عنصر} other{عناصر}}",
   "diveLog_detail_errorLoading": "خطأ في تحميل الغوصة",
-  "diveLog_detail_fullscreen_sampleData": "بيانات العينة",
-  "diveLog_detail_fullscreen_tapChartCompact": "اضغط على المخطط للعرض المدمج",
-  "diveLog_detail_fullscreen_tapChartFull": "اضغط على المخطط للعرض بملء الشاشة",
-  "diveLog_detail_fullscreen_touchChart": "المس المخطط لرؤية البيانات عند تلك النقطة",
   "diveLog_detail_label_airTemp": "درجة حرارة الهواء",
   "diveLog_detail_label_avgDepth": "متوسط العمق",
   "diveLog_detail_label_buddy": "زميل الغوص",
