+++ lib/l10n/arb/app_de.arb
@@ -1116,10 +1116,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{Gegenstand} other{Gegenstaende}}",
   "diveLog_detail_errorLoading": "Fehler beim Laden des Tauchgangs",
-  "diveLog_detail_fullscreen_sampleData": "Messdaten",
-  "diveLog_detail_fullscreen_tapChartCompact": "Tippen Sie auf das Diagramm für kompakte Ansicht",
-  "diveLog_detail_fullscreen_tapChartFull": "Tippen Sie auf das Diagramm für Vollbildansicht",
-  "diveLog_detail_fullscreen_touchChart": "Berühren Sie das Diagramm, um Daten an diesem Punkt zu sehen",
   "diveLog_detail_label_airTemp": "Lufttemperatur",
   "diveLog_detail_label_avgDepth": "Durchschn. Tiefe",
   "diveLog_detail_label_buddy": "Tauchpartner",
