+++ lib/l10n/arb/app_en.arb
@@ -1769,10 +1769,6 @@
   },
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{item} other{items}}",
   "diveLog_detail_errorLoading": "Error loading dive",
-  "diveLog_detail_fullscreen_sampleData": "Sample Data",
-  "diveLog_detail_fullscreen_tapChartCompact": "Tap chart for compact view",
-  "diveLog_detail_fullscreen_tapChartFull": "Tap chart for full-screen view",
-  "diveLog_detail_fullscreen_touchChart": "Touch the chart to see data at that point",
   "diveLog_detail_label_airTemp": "Air Temp",
   "diveLog_detail_label_avgDepth": "Avg Depth",
   "diveLog_detail_label_buddy": "Buddy",
