+++ lib/l10n/arb/app_es.arb
@@ -1116,10 +1116,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{elemento} other{elementos}}",
   "diveLog_detail_errorLoading": "Error al cargar la inmersión",
-  "diveLog_detail_fullscreen_sampleData": "Datos de Muestra",
-  "diveLog_detail_fullscreen_tapChartCompact": "Toca el gráfico para vista compacta",
-  "diveLog_detail_fullscreen_tapChartFull": "Toca el gráfico para vista de pantalla completa",
-  "diveLog_detail_fullscreen_touchChart": "Toca el gráfico para ver datos en ese punto",
   "diveLog_detail_label_airTemp": "Temp. del aire",
   "diveLog_detail_label_avgDepth": "Profundidad media",
   "diveLog_detail_label_buddy": "Compañero",
