+++ lib/l10n/arb/app_fr.arb
@@ -1082,10 +1082,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{equipement} other{equipements}}",
   "diveLog_detail_errorLoading": "Erreur lors du chargement de la plongee",
-  "diveLog_detail_fullscreen_sampleData": "Données d'échantillon",
-  "diveLog_detail_fullscreen_tapChartCompact": "Appuyez sur le graphique pour une vue compacte",
-  "diveLog_detail_fullscreen_tapChartFull": "Appuyez sur le graphique pour une vue plein écran",
-  "diveLog_detail_fullscreen_touchChart": "Touchez le graphique pour voir les données à ce point",
   "diveLog_detail_label_airTemp": "Temp air",
   "diveLog_detail_label_avgDepth": "Profondeur moy",
   "diveLog_detail_label_buddy": "Binome",
