+++ lib/l10n/arb/app_he.arb
@@ -1082,10 +1082,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{פריט} other{פריטים}}",
   "diveLog_detail_errorLoading": "שגיאה בטעינת צלילה",
-  "diveLog_detail_fullscreen_sampleData": "נתוני דגימה",
-  "diveLog_detail_fullscreen_tapChartCompact": "לחץ על התרשים לתצוגה קומפקטית",
-  "diveLog_detail_fullscreen_tapChartFull": "לחץ על התרשים לתצוגה במסך מלא",
-  "diveLog_detail_fullscreen_touchChart": "גע בתרשים כדי לראות נתונים באותה נקודה",
   "diveLog_detail_label_airTemp": "טמפ' אוויר",
   "diveLog_detail_label_avgDepth": "עומק ממוצע",
   "diveLog_detail_label_buddy": "שותף",
