+++ lib/l10n/arb/app_hu.arb
@@ -1082,10 +1082,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{targy} other{targy}}",
   "diveLog_detail_errorLoading": "Hiba a merules betoltesekor",
-  "diveLog_detail_fullscreen_sampleData": "Minta adatok",
-  "diveLog_detail_fullscreen_tapChartCompact": "Koppints a grafikonra a kompakt nézetért",
-  "diveLog_detail_fullscreen_tapChartFull": "Koppints a grafikonra a teljes képernyős nézetért",
-  "diveLog_detail_fullscreen_touchChart": "Érintsd meg a grafikont az adott pont adatainak megtekintéséhez",
   "diveLog_detail_label_airTemp": "Levego hom.",
   "diveLog_detail_label_avgDepth": "Atlag melyseg",
   "diveLog_detail_label_buddy": "Buddy",
