+++ lib/l10n/arb/app_it.arb
@@ -1082,10 +1082,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{elemento} other{elementi}}",
   "diveLog_detail_errorLoading": "Errore nel caricamento dell'immersione",
-  "diveLog_detail_fullscreen_sampleData": "Dati Campione",
-  "diveLog_detail_fullscreen_tapChartCompact": "Tocca il grafico per vista compatta",
-  "diveLog_detail_fullscreen_tapChartFull": "Tocca il grafico per vista a schermo intero",
-  "diveLog_detail_fullscreen_touchChart": "Tocca il grafico per vedere i dati in quel punto",
   "diveLog_detail_label_airTemp": "Temp. aria",
   "diveLog_detail_label_avgDepth": "Profondita media",
   "diveLog_detail_label_buddy": "Compagno",
