+++ lib/l10n/arb/app_localizations.dart
@@ -5402,30 +5402,6 @@ abstract class AppLocalizations {
   /// **'Error loading dive'**
   String get diveLog_detail_errorLoading;
 
-  /// No description provided for @diveLog_detail_fullscreen_sampleData.
-  ///
-  /// In en, this message translates to:
-  /// **'Sample Data'**
-  String get diveLog_detail_fullscreen_sampleData;
-
-  /// No description provided for @diveLog_detail_fullscreen_tapChartCompact.
-  ///
-  /// In en, this message translates to:
-  /// **'Tap chart for compact view'**
-  String get diveLog_detail_fullscreen_tapChartCompact;
-
-  /// No description provided for @diveLog_detail_fullscreen_tapChartFull.
-  ///
-  /// In en, this message translates to:
-  /// **'Tap chart for full-screen view'**
-  String get diveLog_detail_fullscreen_tapChartFull;
-
-  /// No description provided for @diveLog_detail_fullscreen_touchChart.
-  ///
-  /// In en, this message translates to:
-  /// **'Touch the chart to see data at that point'**
-  String get diveLog_detail_fullscreen_touchChart;
-
   /// No description provided for @diveLog_detail_label_airTemp.
   ///
   /// In en, this message translates to:
