+++ lib/l10n/arb/app_localizations_de.dart
@@ -3232,21 +3232,6 @@ class AppLocalizationsDe extends AppLocalizations {
   String get diveLog_detail_errorLoading => 'Fehler beim Laden des Tauchgangs';
 
   @override
-  String get diveLog_detail_fullscreen_sampleData => 'Messdaten';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartCompact =>
-      'Tippen Sie auf das Diagramm für kompakte Ansicht';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartFull =>
-      'Tippen Sie auf das Diagramm für Vollbildansicht';
-
-  @override
-  String get diveLog_detail_fullscreen_touchChart =>
-      'Berühren Sie das Diagramm, um Daten an diesem Punkt zu sehen';
-
-  @override
   String get diveLog_detail_label_airTemp => 'Lufttemperatur';
 
   @override
