+++ lib/l10n/arb/app_localizations_nl.dart
@@ -3200,21 +3200,6 @@ class AppLocalizationsNl extends AppLocalizations {
   String get diveLog_detail_errorLoading => 'Fout bij laden van duik';
 
   @override
-  String get diveLog_detail_fullscreen_sampleData => 'Meetgegevens';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartCompact =>
-      'Tik op grafiek voor compacte weergave';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartFull =>
-      'Tik op grafiek voor volledig scherm weergave';
-
-  @override
-  String get diveLog_detail_fullscreen_touchChart =>
-      'Raak de grafiek aan om gegevens op dat punt te zien';
-
-  @override
   String get diveLog_detail_label_airTemp => 'Luchttemp.';
 
   @override
