+++ lib/l10n/arb/app_localizations_pt.dart
@@ -3225,21 +3225,6 @@ class AppLocalizationsPt extends AppLocalizations {
   String get diveLog_detail_errorLoading => 'Erro ao carregar mergulho';
 
   @override
-  String get diveLog_detail_fullscreen_sampleData => 'Dados de Amostra';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartCompact =>
-      'Toque no gráfico para visualização compacta';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartFull =>
-      'Toque no gráfico para visualização em tela cheia';
-
-  @override
-  String get diveLog_detail_fullscreen_touchChart =>
-      'Toque no gráfico para ver os dados naquele ponto';
-
-  @override
   String get diveLog_detail_label_airTemp => 'Temp do Ar';
 
   @override
