+++ lib/l10n/arb/app_localizations_zh.dart
@@ -3067,18 +3067,6 @@ class AppLocalizationsZh extends AppLocalizations {
   String get diveLog_detail_errorLoading => '加载潜水出错';
 
   @override
-  String get diveLog_detail_fullscreen_sampleData => '采样数据';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartCompact => '点击图表切换紧凑视图';
-
-  @override
-  String get diveLog_detail_fullscreen_tapChartFull => '点击图表切换全屏视图';
-
-  @override
-  String get diveLog_detail_fullscreen_touchChart => '触摸图表查看该点的数据';
-
-  @override
   String get diveLog_detail_label_airTemp => '气温';
 
   @override
