+++ lib/l10n/arb/app_nl.arb
@@ -1116,10 +1116,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{item} other{items}}",
   "diveLog_detail_errorLoading": "Fout bij laden van duik",
-  "diveLog_detail_fullscreen_sampleData": "Meetgegevens",
-  "diveLog_detail_fullscreen_tapChartCompact": "Tik op grafiek voor compacte weergave",
-  "diveLog_detail_fullscreen_tapChartFull": "Tik op grafiek voor volledig scherm weergave",
-  "diveLog_detail_fullscreen_touchChart": "Raak de grafiek aan om gegevens op dat punt te zien",
   "diveLog_detail_label_airTemp": "Luchttemp.",
   "diveLog_detail_label_avgDepth": "Gem. diepte",
   "diveLog_detail_label_buddy": "Buddy",
