+++ lib/l10n/arb/app_pt.arb
@@ -1116,10 +1116,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 field} other{{count} fields}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{item} other{itens}}",
   "diveLog_detail_errorLoading": "Erro ao carregar mergulho",
-  "diveLog_detail_fullscreen_sampleData": "Dados de Amostra",
-  "diveLog_detail_fullscreen_tapChartCompact": "Toque no gráfico para visualização compacta",
-  "diveLog_detail_fullscreen_tapChartFull": "Toque no gráfico para visualização em tela cheia",
-  "diveLog_detail_fullscreen_touchChart": "Toque no gráfico para ver os dados naquele ponto",
   "diveLog_detail_label_airTemp": "Temp do Ar",
   "diveLog_detail_label_avgDepth": "Prof. Media",
   "diveLog_detail_label_buddy": "Dupla",
