+++ lib/l10n/arb/app_zh.arb
@@ -1216,10 +1216,6 @@
   "diveLog_detail_customFieldCount": "{count, plural, =1{1 个字段} other{{count} 个字段}}",
   "diveLog_detail_equipmentCount": "{count} {count, plural, =1{件装备} other{件装备}}",
   "diveLog_detail_errorLoading": "加载潜水出错",
-  "diveLog_detail_fullscreen_sampleData": "采样数据",
-  "diveLog_detail_fullscreen_tapChartCompact": "点击图表切换紧凑视图",
-  "diveLog_detail_fullscreen_tapChartFull": "点击图表切换全屏视图",
-  "diveLog_detail_fullscreen_touchChart": "触摸图表查看该点的数据",
   "diveLog_detail_label_airTemp": "气温",
   "diveLog_detail_label_avgDepth": "平均深度",
   "diveLog_detail_label_buddy": "潜伴",
